Refactor developer action helpers into modules
